B3.29: return-type inference from observed write width
Wire a per-function pass that observes the widest write to the return
register and publishes `void` / `int` / `long`, threaded through both
the C backend's `pick_return_type` and the annotation channel so both
surfaces share a single source. The lattice-driven `int64_t` fallback
and the stale "default void return; lands with B3.6" placeholder are
retired.

- `dac-cli::lift::infer_return_type` walks SSA writes to the return
  register variable AND probes the Return value's source op so a
  function whose Move-into-rax was folded by B3.26's simplifier still
  publishes the right width.
- `RecoveryFacts.inferred_return: InferredReturn`; backend consumes via
  new `Recovered.inferred_return: Option<&CType>` channel. Precedence:
  canonical (B3.28) > inferred (B3.29) > lattice > int64_t fallback.
- Functions whose body returns a value but show no observed writes stay
  at `long` rather than collapsing to `void` (I-6 — avoid breaking the
  round-trip compile gate with `return v<id>;` against a void
  declarator).
- 13 new tests across dac-backend-c, dac-cli::lift, dac-cli::annotations.
- 4 goldens refreshed: int64_t -> long for non-main functions; PE
  `__main` correctly tightens to `int __main(void)` via the post-
  simplifier Load-width-4 probe.

Closes B3.29, FR-18, FR-21, I-3.

+++ tests/golden/hello-elf-o1-c/source.c
@@ -25,7 +25,7 @@ extern int64_t write(int32_t fd, void * buf, uint64_t n);
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
+long _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
     int64_t v0 = 0LL;
     int64_t v2 = 0LL;
     int8_t v5 = 0LL;
@@ -45,7 +45,7 @@ int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t ar
         v12 = ((long long (*)(long long, long long, long long, long long, long long, long long))(/* opaque: indirect-call */ 0))(v6, v7, v8, v9, v10, v11);
     }
     /* phi v13 <- (bb0: v2) (bb1: v12) */
-    return v13;
+    return ((long)(v13));
 }
 
 /* dac-recovered function */
@@ -133,7 +133,7 @@ void _start(void) {
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t deregister_tm_clones(void) {
+long deregister_tm_clones(void) {
     int8_t v2 = 0LL;
     int64_t v3 = 0LL;
     int8_t v6 = 0LL;
@@ -143,7 +143,7 @@ int64_t deregister_tm_clones(void) {
     if (v2) {
 L0:;
         /* phi v7 <- (bb0: 16408) (bb1: v3) */
-        return v7;
+        return ((long)(v7));
     } else {
         v3 = (*((int64_t *)(16328LL)));
         v6 = (v3 == 0LL);
@@ -169,7 +169,7 @@ L0:;
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t register_tm_clones(void) {
+long register_tm_clones(void) {
     int64_t v8 = 0LL;
     int8_t v11 = 0LL;
     int64_t v12 = 0LL;
@@ -191,7 +191,7 @@ int64_t register_tm_clones(void) {
 /* stack_locals: 1 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
+long __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
     int8_t v0 = 0LL;
     int8_t v1 = 0LL;
     int64_t v2 = 0LL;
@@ -214,7 +214,7 @@ int64_t __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t 
     v0 = (*((int8_t *)(16408LL)));
     v1 = (v0 != 0LL);
     if (v1) {
-        return v21;
+        return ((long)(v21));
     } else {
         v3 = ((void *)((v2 - 8LL)));
         *((int64_t *)(((int64_t)(v3)))) = v4;
@@ -246,9 +246,9 @@ void frame_dummy(void) {
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t _fini(void) {
+long _fini(void) {
     int64_t v0 = 0LL;
     int64_t v3 = 0LL;
 
-    return v3;
+    return ((long)(v3));
 }
